# Gen/SpringBootDemo.openApi.yaml
openapi: 3.0.0
info:
  description: System to keep track of avail restaurant menu items
  version: 1.0.0
  title: Customer Orders API
  contact:
    email: SwingGuy1024@yahoo.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
# See https://stackoverflow.com/questions/62044747/message-field-is-empty-in-error-response-spring-boot
#servers:
#  - url: http://localhost:27777/demo/
#    description: Basic server

#server:
#  error:
#    include-message: always
#    include-binding-errors: always
paths:
...
  /menuItem/{id}:
    get:
      summary: Gets a menuItem by ID
      description: Gets a MenuItem by its ID.
      operationId: getMenuItem
      parameters:
        - in: path
          name: id
          description: ID of menuItem to find
          required: true
          schema:
            type: integer
            format: int32
      responses:
        200:
          description: Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItemDto'
        404:
          description: NotFound
...
  /admin/menuItem/addOption/{menuItemId}:
    post:
      summary: Add a menuItemOption
      description: |
        Add an existing MenuItemOption to a MenuItem.
      operationId: addMenuItemOption
      parameters:
        - in: path
          name: menuItemId
          required: true
          description: ID of the MenuItem getting the new option
          schema:
            type: integer
            format: int32
      requestBody:
        description: MenuItemOption
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuItemOptionDto'
      responses:
        201:
          description: MenuItemOption added
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Bad Request
        417:
          description: Token Expired
...
security:
  - bearerAuth: [ ]
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    MenuItemDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        itemPrice:
          type: number
          description: Price. (BigDecimal)
        allowedOptions:
          type: array
          items:
            $ref: '#/components/schemas/MenuItemOptionDto'
      required:
        - name
        - itemPrice
    MenuItemOptionDto:
      type: object
      description: Option for a MenuItem
      properties:
        name:
          type: string
        deltaPrice:
          type: number
          description: Floating point price. Strings are easier to work with.
        id:
          type: integer
          format: int32
      required:
        - name
        - deltaPrice
...